requirements.txt
pandas
numpy
PyYaml
python-dotenv
pymysql
joblib
scikit-learn
nltk
-e .